+++ RPG-System/RPG/Movement.h
@@ -12,7 +12,7 @@ class GameObject;
 class Movement
 {
 public:
-	Movement(sf::Vector2f pos = sf::Vector2f(0.f, 0.f), unsigned speed = DEFAULT_SPEED) : 
+	Movement(sf::Vector2f pos = sf::Vector2f(0.f, 0.f), float speed = DEFAULT_SPEED) : 
 	  _direction(0.f,0.f), _speed(speed), _newpos(true), _init_pos(pos) { }
 	virtual ~Movement() { }
 
@@ -23,7 +23,7 @@ public:
 
 protected:
 	sf::Vector2f _direction;
-	unsigned _speed;
+	float _speed;
 	sf::Vector2f _init_pos;
 	bool _newpos;
 };
