+++ RPG-System/RPG/RandMovement.h
@@ -1,3 +1,5 @@
+// random movement
+
 #pragma once
 
 #include "Movement.h"
@@ -5,7 +7,7 @@
 class RandMovement : public Movement
 {
 public:
-	RandMovement(float time_between_moves, sf::Vector2f pos = sf::Vector2f(0.f, 0.f), unsigned speed = DEFAULT_SPEED) : Movement(pos, speed), _next_move_timer(5.f), 
+	RandMovement(float time_between_moves, sf::Vector2f pos = sf::Vector2f(0.f, 0.f), float speed = DEFAULT_SPEED) : Movement(pos, speed), _next_move_timer(5.f), 
 		_time_between_moves(time_between_moves), _moving_duration_timer(0.f) { }
 	virtual void handleEvents(const Control& controls);
 	virtual void Update(LocalMap& localmap, GameObject& my_obj, Graphics& my_graphics, float elapsedTime);
